Bot.Py/requirements.txt
MetaTrader5
pandas
ta
python-telegram-bot==20.3
pytz
python-dotenv